Rename LayoutComponents to use jsx extension
