+++ lib/LayoutComponents/index.js
@@ -1,9 +1,9 @@
 
 module.exports = {
 
-	_HeaderTitle : require('./_HeaderTitle'),
-	_HeaderSearch : require('./_HeaderSearch'),
-	_HeaderTabs : require('./_HeaderTabs'),
-	_HeaderWaterfall : require('./_HeaderWaterfall'),
+	_HeaderTitle : require('./_HeaderTitle.jsx'),
+	_HeaderSearch : require('./_HeaderSearch.jsx'),
+	_HeaderTabs : require('./_HeaderTabs.jsx'),
+	_HeaderWaterfall : require('./_HeaderWaterfall.jsx'),
 
 };
